Reorganize structs in different files
