+++ src/chart.rs
@@ -1,0 +1,19 @@
+use serde::Deserialize;
+
+/// A representation of a chart definition in a repo.
+#[derive(Debug, Deserialize)]
+pub struct Chart {
+    /// The chart name
+    name: String,
+    /// The chart version
+    version: String,
+}
+
+impl Chart {
+    pub fn version(&self) -> &str {
+        &self.version
+    }
+    pub fn name(&self) -> &str {
+        &self.name
+    }
+}
